Crate imagesize

Structs

ImageSize

Holds the size information of an image.

Enums

ImageError

An Error type used in failure cases.

ImageType

Types of image formats that this crate can identify.

Functions

blob_size

Get the image size from a block of raw data.

image_type

Get the image type from a header

reader_size

Get the image size from a reader

size

Get the image size from a local file

Type Aliases

ImageResult