imagesize

Module util

Enums

Endian

Used for TIFF decoding

Functions

read_bits

read_i32

read_line_capped

read_null_terminated_string

read_tag

Assumes tags are in format of 4 char string followed by big endian size for tag

read_u8

read_u16

read_u24

read_u32

read_u64

read_until_capped

read_until_whitespace

Skips all starting whitespace characters and then reads a string until the next whitespace character Example: “ test string“ => “test”